* MapRefactoring:
  fix: useCascadedShadowMaps: 0
  finish map refactorings
  refactor: cleanup. play session state handling.
  fix: various map bugs, cleanups.
  refactoring map UI interaction. cleaning up.
  fix: finger stage and player pin now work correctly fix: stage does not change player location until you click on a new pin refactor: some cleanups
  update look-at, player locations, journey pos update.
  feat: huge map refactor. feat: ropes are now dynamically joined
  Added ReplaceGameObjectWithPrefab. Map refactoring.-
  refactor PlayerPin, MapPin, Dot.
  refactoring StageMapManager.
  reorganized stage map colors and camera pivots
  Started map refactoring. Reworking StageMapsManager. (Map is broken here!)

+++ Assets/_app/_scripts/Map/PlaySessionState.cs
@@ -3,9 +3,9 @@
     public class PlaySessionState
     {
         public Database.PlaySessionData data;
-        public float score;
+        public int score;
 
-        public PlaySessionState(Database.PlaySessionData _data, float _score)
+        public PlaySessionState(Database.PlaySessionData _data, int _score)
         {
             this.data = _data;
             this.score = _score;
